Rename get_slack_team_members.py to get_slack_usergroup_members.py
Also update imports and entrypoints
